+++ src/main/java/classfit/example/classfit/scoreReport/service/ScoreReportService.java
@@ -19,7 +19,6 @@ import classfit.example.classfit.exam.repository.ExamRepository;
 import classfit.example.classfit.exam.repository.ExamScoreRepository;
 import classfit.example.classfit.member.domain.Member;
 import classfit.example.classfit.scoreReport.domain.ScoreReport;
-import classfit.example.classfit.scoreReport.domain.ScoreReportRepository;
 import classfit.example.classfit.scoreReport.dto.process.ReportExam;
 import classfit.example.classfit.scoreReport.dto.process.StudentList;
 import classfit.example.classfit.scoreReport.dto.request.CreateReportRequest;
@@ -29,6 +28,7 @@ import classfit.example.classfit.scoreReport.dto.response.FindClassStudent;
 import classfit.example.classfit.scoreReport.dto.response.FindReportResponse;
 import classfit.example.classfit.scoreReport.dto.response.SentStudentOpinionResponse;
 import classfit.example.classfit.scoreReport.dto.response.ShowStudentReportResponse;
+import classfit.example.classfit.scoreReport.repository.ScoreReportRepository;
 import classfit.example.classfit.student.domain.Enrollment;
 import classfit.example.classfit.student.domain.Student;
 import classfit.example.classfit.student.repository.EnrollmentRepository;
